+++ packages/forms/examples/src/forms/forms.routes.ts
@@ -24,41 +24,52 @@ export const FORMS_EXAMPLES_ROUTES: Routes = [
 				path: 'autocomplete',
 				pathMatch: 'full',
 				component: FormsAutocompleteDemoPageComponent,
+				data: { meta: { title: 'Autocomplete' } },
+
 			},
 			{
 				path: 'datepicker',
 				pathMatch: 'full',
 				component: FormsDatepickerDemoPageComponent,
+				data: { meta: { title: 'Datepicker' } },
+
 			},
 			{
 				path: 'mask',
 				pathMatch: 'full',
 				component: FormsMaskDemoPageComponent,
+				data: { meta: { title: 'Mask' } },
+
 			},
 			{
 				path: 'range-slider',
 				pathMatch: 'full',
 				component: FormsRangeSliderDemoPageComponent,
+				data: { meta: { title: 'Range Slider' } },
 			},
 			{
 				path: 'search-filter',
 				pathMatch: 'full',
 				component: FormsSearchFilterDemoPageComponent,
+				data: { meta: { title: 'Search Filter' } },
 			},
 			{
 				path: 'timepicker',
 				pathMatch: 'full',
 				component: FormsTimepickerDemoPageComponent,
+				data: { meta: { title: 'Timepicker' } },
 			},
 			{
 				path: 'upload',
 				pathMatch: 'full',
 				component: FormsUploadDemoPageComponent,
+				data: { meta: { title: 'Upload' } },
 			},
 			{
 				path: 'wysiwyg',
 				pathMatch: 'full',
 				component: FormsWysiwygDemoPageComponent,
+				data: { meta: { title: 'WYSIWYG' } },
 			},
 		],
 	},