+++ packages/navigation-menu/examples/src/navigation-menu/index.ts
@@ -1,0 +1,2 @@
+export { NavigationMenuExamplesModule } from './navigation-menu.examples.module';
+export { NAVIGATION_MENU_EXAMPLES_ROUTES } from './navigation-menu.routes';